requirements.txt
Unidecode
beautifulsoup4
pydantic
markdownify
openai
python-dotenv
urllib3
Flask
jsonify
gunicorn
gspread
google-auth
pymongo
oauth2client
seleniumbase